pyproject.toml
[project]
name = "reddit-story"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "crewai>=1.12.2",
    "django>=5.2.12",
    "django-environ>=0.13.0",
    "google-genai>=1.69.0",
    "edge-tts>=7.0.2",
    "gtts>=2.5.4",
    "ollama>=0.6.1",
    "pillow>=12.1.1",
    "sentencepiece>=0.2.1",
    "torch>=2.11.0",
    "transformers>=5.4.0",
]
